Move Bcd type to lib
